+++ packages/auto/utils/changelog.md
@@ -1,3 +1,7 @@
+## v0.3.0
+
+* 🌱 parse git commit descriptions and use it in changelog
+
 ## v0.2.0
 
 * 💥 remove all "repo" methods