+++ packages/primitives/image/changelog.md
@@ -1,3 +1,11 @@
+## v0.4.2
+
+* 🐞 update package.json for image primitive and themeable (#207)
+
+## v0.4.1
+
+* 🐞 move style related props from primitive to themeable for image (#206)
+
 ## v0.4.0
 
 * 🌱 link the style types to themeables/image