+++ packages/primitives/link/changelog.md
@@ -1,3 +1,11 @@
+## v0.9.0
+
+* 🌱 add strong types to `rel` and `target`
+
+## v0.8.0
+
+* 🌱 add support to `rel` attribute on `<a>`
+
 ## v0.7.1
 
 * 🐞 bump react-native