+++ packages/primitives/svg/changelog.md
@@ -1,3 +1,7 @@
+## v0.1.7
+
+* 🐞 bump `react-native-svg` to v10
+
 ## v0.1.6
 
 * 🐞 fix package entrypoints