+++ packages/rebox/ios/changelog.md
@@ -1,3 +1,12 @@
+## v0.8.1
+
+* 🐞 fix `react-native-svg` dependency owner
+
+## v0.8.0
+
+* 💥 upgrade to React Native v0.61
+* ♻️ update dependencies: `@rebox/native-utils`
+
 ## v0.7.0
 
 * ♻️ update dependencies: `@rebox/native-utils`