+++ packages/rebox/native-utils/changelog.md
@@ -1,3 +1,13 @@
+## v0.4.1
+
+* 🐞 fix `react-native-svg` dependency owner
+
+## v0.4.0
+
+* 💥 upgrade to Haul v0.15
+* 💥 upgrade to React Native v0.61
+* 🐞 bump `metro-react-native-babel-preset`
+
 ## v0.3.1
 
 * 🐞 upgrade `is-port-reachable` dependency