+++ packages/start/plugin-lib-istanbul/changelog.md
@@ -1,3 +1,7 @@
+## v0.8.0
+
+* 💥 remove `options` argument from `istanbulInstrument()`
+
 ## v0.7.0
 
 * ♻️ update dependencies: `tsfn`