+++ packages/themeables/image/changelog.md
@@ -1,3 +1,11 @@
+## v0.1.2
+
+* 🐞 update package.json for image primitive and themeable (#207)
+
+## v0.1.1
+
+* 🐞 move style related props from primitive to themeable for image (#206)
+
 ## v0.1.0
 
 * 🐣 init